Changed to np.int16

+++ src/pyate/combo_basic.py
@@ -40,7 +40,7 @@ def combo_basic(
         )
 
     if len(technical_counts) == 0:
-        return pd.Series()
+        return pd.Series(dtype=TermExtraction.dtype)
 
     order = sorted(
         list(technical_counts.keys()), key=TermExtraction.word_length, reverse=True
@@ -52,7 +52,7 @@ def combo_basic(
     technical_counts = technical_counts[order]
 
     if len(technical_counts) == 0:
-        return pd.Series()
+        return pd.Series(dtype=TermExtraction.dtype)
 
     df = pd.DataFrame(
         {
